feat(routes): restructure routing by creating separate Admin and Auth routes with dynamic loading
