+++ resources/ts/App.tsx
@@ -1,4 +1,4 @@
-import AppRoutes from '@/routes/AppRoutes';
+import AppRoutes from '@/AppRoutes';
 
 export default function App() {
   return <AppRoutes />;
