+++ resources/ts/routes/routesConfig.ts
@@ -1,0 +1,6 @@
+import AuthRoutes from "./AuthRoutes";
+import AdminRoutes from "./AdminRoutes";
+
+const routesConfig = [...AuthRoutes, ...AdminRoutes];
+
+export default routesConfig;
